@@ -1,17 +1,17 @@
 import 'package:dartz/dartz.dart';
-import 'package:withconi/configs/constants/enum.dart';
+import 'package:withconi/data/enums/enum.dart';
 import 'package:withconi/controller/auth_controller.dart';
 import 'package:withconi/controller/ui_interpreter/failure_ui_interpreter.dart';
 import 'package:withconi/core/error_handling/failures.dart';
 import 'package:withconi/data/model/user.dart';
 import 'package:withconi/data/repository/user_repository.dart';
-import 'package:withconi/ui/widgets/loading/loading_overlay.dart';
-import 'package:withconi/ui/widgets/photo_gallary/image_item.dart';
-import '../../configs/constants/regex.dart';
-import '../../configs/constants/strings.dart';
-import '../../configs/helpers/image_picker_helper.dart';
+import 'package:withconi/module/widgets/loading/loading_overlay.dart';
+import 'package:withconi/module/widgets/photo_gallary/image_item.dart';
+import '../../core/values/constants/regex.dart';
+import '../../core/values/constants/strings.dart';
+import '../../core/tools/helpers/image_picker_helper.dart';
 import '../../import_basic.dart';
-import '../../ui/widgets/dialog/selection_dialog.dart';
+import '../../module/widgets/dialog/selection_dialog.dart';
 
 class EditUserController extends GetxController {
   final UserRepository _userRepository = UserRepository();
@@ -32,7 +32,7 @@ class EditUserController extends GetxController {
   TextEditingController nameTextController = TextEditingController();
   TextEditingController nickNameTextController = TextEditingController();
   TextEditingController emailTextController = TextEditingController();
-  String get userProvider => providerToValue(_wcUser.provider);
+  String get userProvider => _wcUser.provider.code;
   Image? get providerIcon => Image.asset(
         'assets/icons/$userProvider.png',
         width: 39,
