@@ -1,11 +1,11 @@
 import 'package:dartz/dartz.dart';
-import 'package:withconi/configs/helpers/calculator.dart';
+import 'package:withconi/core/tools/helpers/calculator.dart';
 import 'package:withconi/controller/auth_controller.dart';
 import 'package:withconi/controller/community/community_post_detail_controller.dart';
 import 'package:withconi/controller/ui_interpreter/failure_ui_interpreter.dart';
 import 'package:withconi/data/repository/community_repository.dart';
-import 'package:withconi/ui/widgets/loading/loading_overlay.dart';
-import '../../configs/constants/enum.dart';
+import 'package:withconi/module/widgets/loading/loading_overlay.dart';
+import '../../data/enums/enum.dart';
 import '../../core/error_handling/failures.dart';
 import '../../data/model/post.dart';
 import '../../data/model/report.dart';
@@ -243,7 +243,7 @@ class CommunityPostListController extends GetxController {
 
   void onSortTypeChanged(String? sortingType) {
     if (sortingType != null) {
-      selectedSortType.value = koreanToSortType(sortingType);
+      selectedSortType.value = SortType.getByCode(sortingType);
     }
     resetPage(
         sortType: selectedSortType.value, postType: selectedPostType.value);
