@@ -1,4 +1,4 @@
-import 'package:withconi/configs/helpers/calculator.dart';
+import 'package:withconi/core/tools/helpers/calculator.dart';
 import 'package:withconi/controller/auth_controller.dart';
 import 'package:withconi/controller/infinite_scroll_controller.dart';
 import 'package:withconi/controller/ui_interpreter/failure_ui_interpreter.dart';
@@ -6,14 +6,14 @@ import 'package:withconi/data/model/abstract_class/place_preview.dart';
 import 'package:withconi/data/model/hospital_preview.dart';
 import 'package:withconi/data/model/pharmacy_preview.dart';
 import 'package:withconi/data/repository/community_repository.dart';
-import 'package:withconi/ui/entities/custom_marker.dart';
-import 'package:withconi/ui/entities/location.dart';
-import 'package:withconi/ui/widgets/loading/loading_overlay.dart';
+import 'package:withconi/module/ui_model/custom_marker.dart';
+import 'package:withconi/module/ui_model/location.dart';
+import 'package:withconi/module/widgets/loading/loading_overlay.dart';
 
-import '../../configs/constants/enum.dart';
+import '../../data/enums/enum.dart';
 import '../../data/model/post.dart';
 import '../../import_basic.dart';
-import '../../ui/widgets/photo_gallary/image_item.dart';
+import '../../module/widgets/photo_gallary/image_item.dart';
 import '../ui_helper/infinite_scroll.dart';
 
 class MapMyBookmarkController extends GetxController {
